Move bulk-download-related API tests
